新增房屋增加  是否已发布验证、修改验证登记房屋是否存在入参类型Integer->Long

+++ common/src/main/java/com/jsy/community/qo/lease/HouseLeaseQO.java
@@ -39,12 +39,12 @@ public class HouseLeaseQO implements Serializable {
     @Range(groups = {addLeaseSaleHouse.class}, min = 1 , message = "社区id不正确")
     @NotNull(groups = {addLeaseSaleHouse.class}, message = "社区id不能为空!")
     @ApiModelProperty(value = "社区id")
-    private Integer houseCommunityId;
+    private Long houseCommunityId;
 
     @Range(groups = {addLeaseSaleHouse.class}, min = 1 , message = "房屋id不正确")
     @NotNull(groups = {addLeaseSaleHouse.class}, message = "房屋id不能为空!")
     @ApiModelProperty(value = "房源id")
-    private Integer houseId;
+    private Long houseId;
 
     @ApiModelProperty(value = "房屋租售标题")
     @Length(groups = {addLeaseSaleHouse.class}, min = 1, max = BusinessConst.HOUSE_TITLE_CHAR_MAX, message = "标题长度在1~32之间")
